backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI & LangChain
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langgraph==0.2.16
langchain-ollama==0.1.3
langchain-groq==0.1.9
langchain-huggingface==0.1.2
sentence-transformers
chromadb
redis==5.0.8

# Data Validation
pydantic==2.8.2
pydantic-settings==2.3.4

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1

# CORS & Security
python-jose[cryptography]==3.3.0

# Monitoring & Logging
python-json-logger==2.0.7

